fix: second round of Copilot PR review fixes (v0.0.446)
- ttnn_pipeline: hoist build_tlist outside frame loop (no-op per frame)
- ttnn_pipeline: seed per-frame perturbation noise via generator for reproducibility
- pipeline.py: default torch_dtype float16 -> float32 for CPU compatibility
- terminalCommands: narrow Phase 2 grep filter, drop bare 'warning' match
- option_a_diff.md: replace deleted generate_with_sd35.py references

+++ content/projects/animatediff/animatediff_ttnn/pipeline.py
@@ -17,7 +17,7 @@ from diffusers import AnimateDiffPipeline, DDIMScheduler, MotionAdapter
 def create_animatediff_pipeline(
     model_id: str = "CompVis/stable-diffusion-v1-4",
     adapter_id: str = "guoyww/animatediff-motion-adapter-v1-5-2",
-    torch_dtype: torch.dtype = torch.float16,
+    torch_dtype: torch.dtype = torch.float32,
 ) -> AnimateDiffPipeline:
     """Create a diffusers AnimateDiffPipeline with MotionAdapter.
